+++ CHANGELOG.rst
@@ -196,4 +196,5 @@ Changed / Deprecated / Removed / Fixed / Security
 .. _0.9: https://github.com/matthiask/feincms3/compare/0.8...0.9
 .. _0.10: https://github.com/matthiask/feincms3/compare/0.9...0.10
 .. _0.11: https://github.com/matthiask/feincms3/compare/0.10...0.11
-.. _Next version: https://github.com/matthiask/feincms3/compare/0.11...master
+.. _0.12: https://github.com/matthiask/feincms3/compare/0.11...0.12
+.. _Next version: https://github.com/matthiask/feincms3/compare/0.12...master